changed return type of 'find_larger' & 'find_smaller' to 'long'

+++ src/stack/stack.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   stack.h                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tjun-yu <tjun-yu@student.42.fr>            +#+  +:+       +#+        */
+/*   By: we <we@student.42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/11 14:04:38 by tjun-yu           #+#    #+#             */
-/*   Updated: 2024/02/21 14:52:40 by tjun-yu          ###   ########.fr       */
+/*   Updated: 2024/02/23 16:19:59 by we               ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,8 +39,8 @@ void	rotate_up(t_stack *stack);
 void	rotate_down(t_stack *stack);
 
 int		find_mid(t_stack *stack, int i, int j);
-int		find_larger(t_stack *stack, int target);
-int		find_smaller(t_stack *stack, int target);
+long	find_larger(t_stack *stack, int target);
+long	find_smaller(t_stack *stack, int target);
 int		find_unsorted(t_stack *stack);
 int		find_digit(t_stack *stack, int digit);
 int		find_negative(t_stack *stack);
